refactor(registers/new): utiliza novos steps na criacao de registro
