+++ src/views/pages/dashboard/registers/new/components/CreationRegisterCard/index.tsx
